+++ R/miceFast.R
@@ -1,1 +1,1 @@
-#miceFast
+# miceFast