+++ tests/testthat/test-naive_fill_NA.R
@@ -7,7 +7,7 @@ test_that("auto_imputes", {
 
   test <- sum(is.na(naive_fill_NA(air_miss))) == 0
 
-  test_1 <- identical(vapply(naive_fill_NA(air_miss),class,character(1)),vapply(air_miss,class,character(1)))
+  test_1 <- identical(vapply(naive_fill_NA(air_miss), class, character(1)), vapply(air_miss, class, character(1)))
 
-  test_all <- expect_true(all(c(test,test_1)))
+  test_all <- expect_true(all(c(test, test_1)))
 })